moved functions to utility.php
