+++ blocks/wflinks_block_tag.php
@@ -8,8 +8,8 @@
 function wflinks_tag_block_cloud_show($options)
 {
     $moduleDirName = basename(dirname(__DIR__));
-    require_once XOOPS_ROOT_PATH . '/modules/' . $moduleDirName . '/include/functions.php';
-    if (wfl_tag_module_included()) {
+    require_once XOOPS_ROOT_PATH . '/modules/' . $moduleDirName . '/class/utility.php';
+    if (WfLinksUtility::isTagModuleIncluded()) {
         require_once XOOPS_ROOT_PATH . '/modules/tag/blocks/block.php';
 
         return tag_block_cloud_show($options, $moduleDirName);
@@ -26,8 +26,8 @@ function wflinks_tag_block_cloud_show($options)
 function wflinks_tag_block_cloud_edit($options)
 {
     $moduleDirName = basename(dirname(__DIR__));
-    require_once XOOPS_ROOT_PATH . '/modules/' . $moduleDirName . '/include/functions.php';
-    if (wfl_tag_module_included()) {
+    require_once XOOPS_ROOT_PATH . '/modules/' . $moduleDirName . '/class/utility.php';
+    if (WfLinksUtility::isTagModuleIncluded()) {
         require_once XOOPS_ROOT_PATH . '/modules/tag/blocks/block.php';
 
         return tag_block_cloud_edit($options);
@@ -44,8 +44,8 @@ function wflinks_tag_block_cloud_edit($options)
 function wflinks_tag_block_top_show($options)
 {
     $moduleDirName = basename(dirname(__DIR__));
-    require_once XOOPS_ROOT_PATH . '/modules/' . $moduleDirName . '/include/functions.php';
-    if (wfl_tag_module_included()) {
+    require_once XOOPS_ROOT_PATH . '/modules/' . $moduleDirName . '/class/utility.php';
+    if (WfLinksUtility::isTagModuleIncluded()) {
         require_once XOOPS_ROOT_PATH . '/modules/tag/blocks/block.php';
 
         return tag_block_top_show($options, $moduleDirName);
@@ -62,8 +62,8 @@ function wflinks_tag_block_top_show($options)
 function wflinks_tag_block_top_edit($options)
 {
     $moduleDirName = basename(dirname(__DIR__));
-    require_once XOOPS_ROOT_PATH . '/modules/' . $moduleDirName . '/include/functions.php';
-    if (wfl_tag_module_included()) {
+    require_once XOOPS_ROOT_PATH . '/modules/' . $moduleDirName . '/class/utility.php';
+    if (WfLinksUtility::isTagModuleIncluded()) {
         require_once XOOPS_ROOT_PATH . '/modules/tag/blocks/block.php';
 
         return tag_block_top_edit($options);
