+++ class/common/traitfilesmgmt.php
@@ -1,0 +1,90 @@
+<?php
+/*
+ You may not change or alter any portion of this comment or credits
+ of supporting developers from this source code or any supporting source code
+ which is considered copyrighted (c) material of the original comment or credit authors.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ */
+/**
+ * @copyright   XOOPS Project (https://xoops.org)
+ * @license     http://www.fsf.org/copyleft/gpl.html GNU public license
+ * @author      mamba <mambax7@gmail.com>
+ */
+
+trait FilesManagement
+{
+    /**
+     * Function responsible for checking if a directory exists, we can also write in and create an index.html file
+     *
+     * @param string $folder The full path of the directory to check
+     *
+     * @return void
+     */
+    public static function createFolder($folder)
+    {
+        //        try {
+        //            if (!mkdir($folder) && !is_dir($folder)) {
+        //                throw new \RuntimeException(sprintf('Unable to create the %s directory', $folder));
+        //            } else {
+        //                file_put_contents($folder . '/index.html', '<script>history.go(-1);</script>');
+        //            }
+        //        }
+        //        catch (Exception $e) {
+        //            echo 'Caught exception: ', $e->getMessage(), "\n", '<br>';
+        //        }
+        try {
+            if (!file_exists($folder)) {
+                if (!mkdir($folder) && !is_dir($folder)) {
+                    throw new \RuntimeException(sprintf('Unable to create the %s directory', $folder));
+                }
+
+                file_put_contents($folder . '/index.html', '<script>history.go(-1);</script>');
+            }
+        } catch (Exception $e) {
+            echo 'Caught exception: ', $e->getMessage(), "\n", '<br>';
+        }
+    }
+
+    /**
+     * @param $file
+     * @param $folder
+     * @return bool
+     */
+    public static function copyFile($file, $folder)
+    {
+        return copy($file, $folder);
+        //        try {
+        //            if (!is_dir($folder)) {
+        //                throw new \RuntimeException(sprintf('Unable to copy file as: %s ', $folder));
+        //            } else {
+        //                return copy($file, $folder);
+        //            }
+        //        } catch (Exception $e) {
+        //            echo 'Caught exception: ', $e->getMessage(), "\n", "<br>";
+        //        }
+        //        return false;
+    }
+
+    /**
+     * @param $src
+     * @param $dst
+     */
+    public static function recurseCopy($src, $dst)
+    {
+        $dir = opendir($src);
+        //    @mkdir($dst);
+        while (false !== ($file = readdir($dir))) {
+            if (($file !== '.') && ($file !== '..')) {
+                if (is_dir($src . '/' . $file)) {
+                    self::recurseCopy($src . '/' . $file, $dst . '/' . $file);
+                } else {
+                    copy($src . '/' . $file, $dst . '/' . $file);
+                }
+            }
+        }
+        closedir($dir);
+    }
+}
