+++ singlelink.php
@@ -13,8 +13,8 @@ require_once __DIR__ . '/header.php';
 
 global $xoopsTpl, $xoTheme;
 
-$lid = (int)wfl_cleanRequestVars($_REQUEST, 'lid', 0);
-$cid = (int)wfl_cleanRequestVars($_REQUEST, 'cid', 0);
+$lid = (int)WfLinksUtility::cleanRequestVars($_REQUEST, 'lid', 0);
+$cid = (int)WfLinksUtility::cleanRequestVars($_REQUEST, 'cid', 0);
 
 $sql2 = 'SELECT count(*) FROM '
         . $xoopsDB->prefix('wflinks_links')
@@ -34,7 +34,7 @@ $sql2 = 'SELECT count(*) FROM '
         . '))';
 list($count) = $xoopsDB->fetchRow($xoopsDB->query($sql2));
 
-if (false === wfl_checkgroups($cid = 0) && $count == 0) {
+if (false === WfLinksUtility::checkGroups($cid = 0) && $count == 0) {
     redirect_header('index.php', 1, _MD_WFL_MUSTREGFIRST);
 }
 
@@ -57,12 +57,12 @@ require_once XOOPS_ROOT_PATH . '/modules/' . $xoopsModule->getVar('dirname') . '
 require_once XOOPS_ROOT_PATH . '/modules/' . $xoopsModule->getVar('dirname') . '/include/address.php';
 
 // tags support
-if (wfl_tag_module_included()) {
+if (WfLinksUtility::isTagModuleIncluded()) {
     require_once XOOPS_ROOT_PATH . '/modules/tag/include/tagbar.php';
     $xoopsTpl->assign('tagbar', tagBar($link_arr['lid'], 0));
 }
 
-//$link['imageheader'] = wfl_imageheader();
+//$link['imageheader'] = WfLinksUtility::getImageHeader();
 $link['id']           = $link_arr['lid'];
 $link['cid']          = $link_arr['cid'];
 $link['description2'] = $wfmyts->displayTarea($link_arr['description'], 1, 1, 1, 1, 1);
@@ -80,7 +80,7 @@ $voip    = $link_arr['voip'];
 $fax     = $link_arr['fax'];
 $email   = $link_arr['email'];
 $vat     = $link_arr['vat'];
-$country = wfl_countryname($link_arr['country']);
+$country = WfLinksUtility::getCountryName($link_arr['country']);
 
 if ($street1 === '' || $town === '' || $xoopsModuleConfig['useaddress'] == 0) {
     $link['addryn'] = 0;
@@ -150,7 +150,7 @@ $link['path'] = $pathstring;
 // Start of meta tags
 $maxWords = 100;
 $words    = array();
-$words    = explode(' ', wfl_html2text($link_arr['description']));
+$words    = explode(' ', WfLinksUtility::convertHtml2text($link_arr['description']));
 $newWords = array();
 $i        = 0;
