+++ submit.php
@@ -17,26 +17,26 @@ include XOOPS_ROOT_PATH . '/class/xoopsformloader.php';
 $mytree = new WflinksXoopsTree($xoopsDB->prefix('wflinks_cat'), 'cid', 'pid');
 global $wfmyts, $xoopsModuleConfig;
 
-$cid = wfl_cleanRequestVars($_REQUEST, 'cid', 0);
-$lid = wfl_cleanRequestVars($_REQUEST, 'lid', 0);
+$cid = WfLinksUtility::cleanRequestVars($_REQUEST, 'cid', 0);
+$lid = WfLinksUtility::cleanRequestVars($_REQUEST, 'lid', 0);
 $cid = (int)$cid;
 $lid = (int)$lid;
 
-if (false === wfl_checkgroups($cid, 'WFLinkSubPerm')) {
+if (false === WfLinksUtility::checkGroups($cid, 'WFLinkSubPerm')) {
     redirect_header('index.php', 1, _MD_WFL_NOPERMISSIONTOPOST);
 }
 
-if (true === wfl_checkgroups($cid, 'WFLinkSubPerm')) {
-    if (wfl_cleanRequestVars($_REQUEST, 'submit', 0)) {
-        if (false === wfl_checkgroups($cid, 'WFLinkSubPerm')) {
+if (true === WfLinksUtility::checkGroups($cid, 'WFLinkSubPerm')) {
+    if (WfLinksUtility::cleanRequestVars($_REQUEST, 'submit', 0)) {
+        if (false === WfLinksUtility::checkGroups($cid, 'WFLinkSubPerm')) {
             redirect_header('index.php', 1, _MD_WFL_NOPERMISSIONTOPOST);
         }
 
         $submitter    = (is_object($xoopsUser) && !empty($xoopsUser)) ? $xoopsUser->getVar('uid') : 0;
-        $forumid      = wfl_cleanRequestVars($_REQUEST, 'forumid', 0);
-        $offline      = wfl_cleanRequestVars($_REQUEST, 'offline', 0);
-        $notifypub    = wfl_cleanRequestVars($_REQUEST, 'notifypub', 0);
-        $approve      = wfl_cleanRequestVars($_REQUEST, 'approve', 0);
+        $forumid      = WfLinksUtility::cleanRequestVars($_REQUEST, 'forumid', 0);
+        $offline      = WfLinksUtility::cleanRequestVars($_REQUEST, 'offline', 0);
+        $notifypub    = WfLinksUtility::cleanRequestVars($_REQUEST, 'notifypub', 0);
+        $approve      = WfLinksUtility::cleanRequestVars($_REQUEST, 'approve', 0);
         $url          = $wfmyts->addSlashes(ltrim($_POST['url']));
         $title        = $wfmyts->addSlashes(ltrim($_REQUEST['title']));
         $descriptionb = $wfmyts->addSlashes(ltrim($_REQUEST['descriptionb']));
@@ -61,7 +61,7 @@ if (true === wfl_checkgroups($cid, 'WFLinkSubPerm')) {
             $fax       = $wfmyts->addSlashes(ltrim($_REQUEST['fax']));
             $voip      = $wfmyts->addSlashes(ltrim($_REQUEST['voip']));
             $mobile    = $wfmyts->addSlashes(ltrim($_REQUEST['mobile']));
-            $email     = emailcnvrt($wfmyts->addSlashes(ltrim($_REQUEST['email'])));
+            $email     = WfLinksUtility::convertEmail($wfmyts->addSlashes(ltrim($_REQUEST['email'])));
             $vat       = $wfmyts->addSlashes(ltrim($_REQUEST['vat']));
         } else {
             $googlemap = $yahoomap = $multimap = $street1 = $street2 = $town = $state = $zip = $tel = $fax = $voip = $mobile = $email = $vat = '';
@@ -77,7 +77,7 @@ if (true === wfl_checkgroups($cid, 'WFLinkSubPerm')) {
             $status      = 0;
             $publishdate = 0;
             $message     = _MD_WFL_THANKSFORINFO;
-            if (true === wfl_checkgroups($cid, 'WFLinkAutoApp')) {
+            if (true === WfLinksUtility::checkGroups($cid, 'WFLinkAutoApp')) {
                 $publishdate = time();
                 $status      = 1;
                 $message     = _MD_WFL_ISAPPROVED;
@@ -98,9 +98,9 @@ if (true === wfl_checkgroups($cid, 'WFLinkSubPerm')) {
 
             // Add item_tag to Tag-module
             if ($lid == 0) {
-                $tagupdate = wfl_tagupdate($newid, $item_tag);
+                $tagupdate = WfLinksUtility::updateTag($newid, $item_tag);
             } else {
-                $tagupdate = wfl_tagupdate($lid, $item_tag);
+                $tagupdate = WfLinksUtility::updateTag($lid, $item_tag);
             }
 
             // Notify of new link (anywhere) and new link in category
@@ -116,7 +116,7 @@ if (true === wfl_checkgroups($cid, 'WFLinkSubPerm')) {
 
             $tags['CATEGORY_NAME'] = $row['title'];
             $tags['CATEGORY_URL']  = XOOPS_URL . '/modules/' . $xoopsModule->getVar('dirname') . '/viewcat.php?cid=' . $cid;
-            if (true === wfl_checkgroups($cid, 'WFLinkAutoApp')) {
+            if (true === WfLinksUtility::checkGroups($cid, 'WFLinkAutoApp')) {
                 $notificationHandler->triggerEvent('global', 0, 'new_link', $tags);
                 $notificationHandler->triggerEvent('category', $cid, 'new_link', $tags);
                 redirect_header('index.php', 2, _MD_WFL_ISAPPROVED);
@@ -131,7 +131,7 @@ if (true === wfl_checkgroups($cid, 'WFLinkSubPerm')) {
                 redirect_header('index.php', 2, _MD_WFL_THANKSFORINFO);
             }
         } else {
-            if (true === wfl_checkgroups($cid, 'WFLinkAutoApp') || $approve == 1) {
+            if (true === WfLinksUtility::checkGroups($cid, 'WFLinkAutoApp') || $approve == 1) {
                 $updated = time();
                 $sql     = 'UPDATE '
                            . $xoopsDB->prefix('wflinks_links')
@@ -159,7 +159,7 @@ if (true === wfl_checkgroups($cid, 'WFLinkSubPerm')) {
                 $modifysubmitter = $xoopsUser->uid();
                 $requestid       = $modifysubmitter;
                 $requestdate     = time();
-                $updated         = wfl_cleanRequestVars($_REQUEST, 'up_dated', time());
+                $updated         = WfLinksUtility::cleanRequestVars($_REQUEST, 'up_dated', time());
                 $sql             = 'INSERT INTO '
                                    . $xoopsDB->prefix('wflinks_mod')
                                    . ' (requestid, lid, cid, title, url, forumid, description, modifysubmitter, requestdate, country, keywords, item_tag, googlemap, yahoomap, multimap, street1, street2, town, state, zip, tel, fax, voip, mobile, email, vat)';
@@ -188,14 +188,14 @@ if (true === wfl_checkgroups($cid, 'WFLinkSubPerm')) {
     } else {
         global $xoopsModuleConfig;
 
-        $approve = wfl_cleanRequestVars($_REQUEST, 'approve', 0);
+        $approve = WfLinksUtility::cleanRequestVars($_REQUEST, 'approve', 0);
 
         //Show disclaimer
         if ($xoopsModuleConfig['showdisclaimer'] && !isset($_GET['agree']) && $approve == 0) {
             $GLOBALS['xoopsOption']['template_main'] = 'wflinks_disclaimer.tpl';
             include XOOPS_ROOT_PATH . '/header.php';
 
-            $xoopsTpl->assign('image_header', wfl_imageheader());
+            $xoopsTpl->assign('image_header', WfLinksUtility::getImageHeader());
             $xoopsTpl->assign('disclaimer', $wfmyts->displayTarea($xoopsModuleConfig['disclaimer'], 1, 1, 1, 1, 1));
             $xoopsTpl->assign('cancel_location', XOOPS_URL . '/modules/' . $xoopsModule->getVar('dirname') . '/index.php');
             $xoopsTpl->assign('link_disclaimer', false);
@@ -209,7 +209,7 @@ if (true === wfl_checkgroups($cid, 'WFLinkSubPerm')) {
             exit();
         }
 
-        echo "<br><div class='txtcenter;'>" . wfl_imageheader() . "</div><br>\n";
+        echo "<br><div class='txtcenter;'>" . WfLinksUtility::getImageHeader() . "</div><br>\n";
         echo '<div>' . _MD_WFL_SUB_SNEWMNAMEDESC . "</div>\n<br>\n";
 
         $sql        = 'SELECT * FROM ' . $xoopsDB->prefix('wflinks_links') . ' WHERE lid=' . $lid;
@@ -268,7 +268,7 @@ if (true === wfl_checkgroups($cid, 'WFLinkSubPerm')) {
         $sql        = 'SELECT * FROM ' . $xoopsDB->prefix('wflinks_cat') . ' ORDER BY title';
         $result     = $xoopsDB->query($sql);
         while ($myrow = $xoopsDB->fetchArray($result)) {
-            if (true === wfl_checkgroups($myrow['cid'], 'WFLinkSubPerm')) {
+            if (true === WfLinksUtility::checkGroups($myrow['cid'], 'WFLinkSubPerm')) {
                 $submitcats[$myrow['cid']] = $myrow['title'];
             }
         }
@@ -278,7 +278,7 @@ if (true === wfl_checkgroups($cid, 'WFLinkSubPerm')) {
         ob_end_clean();
 
         // Link description form
-        //    $editor = wfl_getWysiwygForm( _MD_WFL_DESCRIPTIONC, 'descriptionb', $descriptionb, 15, 60 );
+        //    $editor = WfLinksUtility::getWysiwygForm( _MD_WFL_DESCRIPTIONC, 'descriptionb', $descriptionb, 15, 60 );
         //    $sform -> addElement($editor,false);
 
         $optionsTrayNote = new XoopsFormElementTray(_MD_WFL_DESCRIPTIONC, '<br>');
@@ -305,7 +305,7 @@ if (true === wfl_checkgroups($cid, 'WFLinkSubPerm')) {
 
         // Insert tags if Tag-module is installed and if user is allowed
         if ($xoopsModuleConfig['usercantag']) {
-            if (wfl_tag_module_included()) {
+            if (WfLinksUtility::isTagModuleIncluded()) {
                 require_once XOOPS_ROOT_PATH . '/modules/tag/include/formtag.php';
                 $text_tags = new TagFormTag('item_tag', 70, 255, $link_array['item_tag'], 0);
                 $sform->addElement($text_tags);
@@ -375,11 +375,11 @@ if (true === wfl_checkgroups($cid, 'WFLinkSubPerm')) {
         } else {
             $sform->addElement(new XoopsFormHidden('notifypub', 0));
         }
-        if (true === wfl_checkgroups($cid, 'WFLinkAppPerm') && $lid > 0) {
+        if (true === WfLinksUtility::checkGroups($cid, 'WFLinkAppPerm') && $lid > 0) {
             $approve_checkbox = new XoopsFormCheckBox('', 'approve', $approve);
             $approve_checkbox->addOption(1, _MD_WFL_APPROVE);
             $option_tray->addElement($approve_checkbox);
-        } elseif (true === wfl_checkgroups($cid, 'WFLinkAutoApp')) {
+        } elseif (true === WfLinksUtility::checkGroups($cid, 'WFLinkAutoApp')) {
             $sform->addElement(new XoopsFormHidden('approve', 1));
         } else {
             $sform->addElement(new XoopsFormHidden('approve', 0));
