+++ visit.php
@@ -13,9 +13,9 @@ require_once __DIR__ . '/header.php';
 
 global $xoopsModuleConfig;
 
-$agreed = wfl_cleanRequestVars($_REQUEST, 'agree', 0);
-$cid    = wfl_cleanRequestVars($_REQUEST, 'cid', 0);
-$lid    = wfl_cleanRequestVars($_REQUEST, 'lid', 0);
+$agreed = WfLinksUtility::cleanRequestVars($_REQUEST, 'agree', 0);
+$cid    = WfLinksUtility::cleanRequestVars($_REQUEST, 'cid', 0);
+$lid    = WfLinksUtility::cleanRequestVars($_REQUEST, 'lid', 0);
 $cid    = (int)$cid;
 $lid    = (int)$lid;
 $agreed = (int)$agreed;
@@ -38,7 +38,7 @@ $sql2 = 'SELECT count(*) FROM '
         . '))';
 list($count) = $xoopsDB->fetchRow($xoopsDB->query($sql2));
 
-if (false === wfl_checkgroups($cid) && $count == 0) {
+if (false === WfLinksUtility::checkGroups($cid) && $count == 0) {
     redirect_header('index.php', 1, _MD_WFL_MUSTREGFIRST);
 }
 
@@ -46,7 +46,7 @@ if ($xoopsModuleConfig['showlinkdisclaimer'] && $agreed == 0) {
     $GLOBALS['xoopsOption']['template_main'] = 'wflinks_disclaimer.tpl';
     include XOOPS_ROOT_PATH . '/header.php';
 
-    $xoopsTpl->assign('image_header', wfl_imageheader());
+    $xoopsTpl->assign('image_header', WfLinksUtility::getImageHeader());
     $xoopsTpl->assign('linkdisclaimer', $wfmyts->displayTarea($xoopsModuleConfig['linkdisclaimer'], 1, 1, 1, 1, 1));
     $xoopsTpl->assign('cancel_location', XOOPS_URL . '/modules/' . $xoopsModule->getVar('dirname') . '/index.php');
     $xoopsTpl->assign('agree_location', XOOPS_URL . '/modules/' . $xoopsModule->getVar('dirname') . '/visit.php?agree=1&amp;lid=' . $lid . '&amp;cid=' . $cid);
@@ -61,7 +61,7 @@ if ($xoopsModuleConfig['showlinkdisclaimer'] && $agreed == 0) {
 
     $sql = 'SELECT url FROM ' . $xoopsDB->prefix('wflinks_links') . ' WHERE lid=' . $lid;
     if (!$result = $xoopsDB->queryF($sql)) {
-        echo "<br><div style='text-align: center;'>" . wfl_imageheader() . '</div>';
+        echo "<br><div style='text-align: center;'>" . WfLinksUtility::getImageHeader() . '</div>';
         reportBroken($lid);
     } else {
         list($url) = $xoopsDB->fetchRow($result);
